requirements_production.txt
# Crypto Trading Bot V3 - Production Dependencies

# ==================== Core ====================
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# ==================== Database ====================
psycopg2-binary>=2.9.9

# ==================== Machine Learning ====================
tensorflow>=2.16.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# ==================== Technical Analysis ====================
pandas_ta>=0.3.14b
ta>=0.10.0

# ==================== Exchange Connectors ====================
alpaca-py>=0.13.0
ccxt>=4.0.0

# ==================== Vision / Charts ====================
mplfinance>=0.12.10b0
matplotlib>=3.8.0
Pillow>=10.0.0

# ==================== Telegram ====================
python-telegram-bot>=21.0

# ==================== HTTP / Async ====================
aiohttp>=3.9.0
requests>=2.31.0

# ==================== Utilities ====================
tqdm>=4.66.0
rich>=13.0.0
pytz>=2024.1
